Catch and warn on getSymbols.yahoo() errors
A long-term annoyance with getSymbols() is that an error downloading
or reading data for any symbol causes no data to be loaded.

Wrap the for loop body inside a try() call, and convert any errors
into warnings. Also keep track of symbols that failed to load, so we
only return ticker symbols that actually have data.

See #135.